MSVC warnings fix (#2620)
* Fix MSVC build warnings in Ngraph

* Fix MSVC build warnings in transformations library

* Fix MSVC build warnings in core,legacy,preprocessing

* Fix MSVC build warnings in XLink

* Fix MSVC build warnings in Myriad plugin

+++ inference-engine/src/legacy_api/src/ie_cnn_layer_builder_ngraph.cpp
@@ -1532,7 +1532,7 @@ CNNLayer::Ptr NodeConverter<ngraph::op::PriorBoxIE>::createLayer(const std::shar
         auto img_H = img_shape[2];
         auto data_H = data_shape[2];
         if (attr.step == -1)
-            attr.step = 1. * img_H / data_H;
+            attr.step = static_cast<float>(1. * img_H / data_H);
         else
             attr.step *= img_H;
         for (auto& size : attr.min_size)
