+++ inference-engine/src/transformations/src/transformations/common_optimizations/depth_to_space_fusion.cpp
@@ -18,7 +18,7 @@ bool check_block_first(const ngraph::Shape& shape_input, const ngraph::Shape& sh
     possible_block_size = shape_reshape_before[1];
     if (possible_block_size == 0)
         return false;
-    uint64_t c_dim = shape_input[1] / std::pow(possible_block_size, spatial_dims);
+    uint64_t c_dim = static_cast<uint64_t>(shape_input[1] / std::pow(possible_block_size, spatial_dims));
 
     // x' = reshape(data, [N, block_size, block_size, ..., block_size, C / (block_size ^ K), D1, D2, ..., DK])
     ngraph::Shape expected_shape = {shape_input[0]};
@@ -54,7 +54,7 @@ bool check_depth_first(const ngraph::Shape& shape_input, const ngraph::Shape& sh
     possible_block_size = shape_reshape_before[2];
     if (possible_block_size == 0)
         return false;
-    uint64_t c_dim = shape_input[1] / std::pow(possible_block_size, spatial_dims);
+    uint64_t c_dim = static_cast<uint64_t>(shape_input[1] / std::pow(possible_block_size, spatial_dims));
 
     // x' = reshape(data, [N, C / (block_size ^ K), block_size, block_size, ..., block_size, D1, D2, ..., DK])
     ngraph::Shape expected_shape = {shape_input[0], static_cast<size_t>(c_dim)};
